Move `state` and `prop` outside the `updated` folder
