+++ README.md
@@ -3,7 +3,8 @@
 A thin wrapper around the [Lit](https://github.com/lit/lit) library consisting of some additional features:
 
 - ### [`Component` class](./packages/Lit/Component/README.md)
-- ### [`updated` Decorator (along with upgraded `state` and `property` decorators)](./packages/Lit/updated/README.md)
+- ### [`updated` Decorator](./packages/Lit/updated/README.md)
 - ### [`event` Decorator](./packages/Lit/event/README.md)
 - ### [`eventListener` Decorator](./packages/Lit/eventListener/README.md)
 - ### [`style` Directive](./packages/Lit/style/README.md)
+- ### [`bind` Directive](./packages/Lit/bind/README.md)
