+++ packages/Lit/index.ts
@@ -36,3 +36,5 @@ export * from './updated/index.js'
 export * from './eventListener/index.js'
 export * from './event/index.js'
 export * from './bind/index.js'
+export * from './property.js'
+export * from './state.js'
