+++ packages/Lit/property.ts
@@ -1,7 +1,7 @@
 import type { PropertyDeclaration, ReactiveElement } from 'lit'
 import { property as litProperty } from 'lit/decorators.js'
-import { updated, UpdatedCallback } from './updated.js'
-import { associatedEvent, bindingDefaultProperty } from '../bind/index.js'
+import { updated, UpdatedCallback } from './updated/updated.js'
+import { associatedEvent, bindingDefaultProperty } from './bind/index.js'
 
 export const property = <T>(options?: PropertyDeclaration & {
 	updated?: UpdatedCallback<T>
