+++ packages/Lit/state.ts
@@ -1,6 +1,6 @@
 import type { ReactiveElement } from 'lit'
 import { InternalPropertyDeclaration, state as litState } from 'lit/decorators.js'
-import { updated, UpdatedCallback } from './updated.js'
+import { updated, UpdatedCallback } from './updated/updated.js'
 
 export const state = <T>(options?: InternalPropertyDeclaration & { updated?: UpdatedCallback<T> }) => {
 	return (prototype: ReactiveElement, propertyKey: PropertyKey) => {
