+++ packages/Lit/updated/index.ts
@@ -1,3 +1,1 @@
 export * from './updated.js'
-export * from './property.js'
-export * from './state.js'
